# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Restaurant QR Ordering API
  version: 0.1.0
  description: >
    HTTP API for QR-based restaurant ordering. Most staff endpoints require a JWT
    bearer token and a server-resolved permission. Guest endpoints under /api/v1/guest require an
    X-Session-Token issued by /api/v1/dining/join-session. Public endpoints require
    neither.
  x-response-envelope:
    description: Every JSON response includes the numeric HTTP status in `code`.
servers:
  - url: /
...
paths:
...
  /api/v1/ordering/place-order:
    post:
      tags: [ordering]
      summary: Place an order (staff)
      security:
        - bearerAuth: []
      description: 'Permission: ordering.operate'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScaffoldRequest'
      responses:
        '200':
          description: Order placed
  /api/v1/ordering/cancel-or-edit-item:
    post:
      tags: [ordering]
      summary: Cancel or edit an order item (staff)
      security:
        - bearerAuth: []
      description: 'Permission: ordering.operate'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScaffoldRequest'
      responses:
        '200':
          description: Item updated
  /api/v1/ordering/update-item-status:
    post:
      tags: [ordering]
      summary: Update an order item's status (staff)
      security:
        - bearerAuth: []
      description: 'Permission: ordering.operate'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScaffoldRequest'
      responses:
        '200':
          description: Status updated
  /api/v1/ordering/review-cancel-request:
    post:
      tags: [ordering]
      summary: Approve or reject a guest cancel request (staff)
      security:
        - bearerAuth: []
      description: 'Permission: ordering.operate'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScaffoldRequest'
      responses:
        '200':
          description: Request reviewed
  /api/v1/staff/tables:
    get:
      tags: [ordering]
      summary: List tables, active sessions, orders, and item statuses for staff
      security:
        - bearerAuth: []
      description: 'Permission: ordering.staff'
      responses:
        '200':
          description: Staff table board
  /api/v1/staff/sessions/{sessionId}/request-bill:
    post:
      tags: [ordering]
      summary: Mark a dining session as requesting the bill
      security:
        - bearerAuth: []
      description: 'Permission: ordering.staff. This action uses only the sessionId path parameter.'
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Bill request recorded
  /api/v1/staff/order-items/{itemId}/status:
    patch:
      tags: [ordering]
      summary: Update an order item status from the staff board
      security:
        - bearerAuth: []
      description: 'Permission: ordering.staff'
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateItemStatusRequest'
      responses:
        '200':
          description: Item status updated
  /api/v1/kitchen/queue:
    get:
      tags: [kitchen]
      summary: Kitchen display queue
      security:
        - bearerAuth: []
      description: 'Permission: kitchen.operate'
      responses:
        '501':
          description: Scaffold stub
  /api/v1/kitchen/items/{id}/status:
    patch:
      tags: [kitchen]
      summary: Update kitchen item status
      security:
        - bearerAuth: []
      description: 'Permission: kitchen.operate'
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateItemStatusRequest'
      responses:
        '501':
          description: Scaffold stub
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    sessionToken:
      type: apiKey
      in: header
      name: X-Session-Token
  schemas:
    ScaffoldRequest:
      type: object
      description: Endpoint scaffold. The backend currently accepts an empty JSON object only.
      additionalProperties: false
      example: {}
...
    UpdateItemStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [PENDING, PREPARING, READY, SERVED, CANCELLED]
...